requirements.txt
requests
python-dotenv
django
djangorestframework
google-generativeai